bagamba/pyproject.toml
[project]
name = "incident-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "slack-sdk>=3.36.0",
    "slack-bolt>=1.18.0",
    "jira>=3.5.0",
    "python-dotenv>=1.0.0",
    "asyncio-mqtt>=0.16.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
    "redis>=5.0.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.181.0",
    "ruff>=0.13.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "black>=25.1.0",
    "fastapi>=0.117.1",
    "uvicorn>=0.37.0",
]

[tool.ruff]
target-version = "py311"
output-format = "grouped"
